pyproject.toml
[project]
name = "a2a-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.16",
    "envyaml>=1.10.211231",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langfuse>=3.2.1",
    "langgraph>=0.5.4",
    "oci>=2.155.2",
    "python-box>=7.3.2",
    "uvicorn>=0.35.0",
]
